+++ AntColony/Core/Ants/IAnt.cs
@@ -1,4 +1,5 @@
 ﻿using AntColony.Core.Graphs;
+using System;
 using System.Collections.Generic;
 
 namespace AntColony.Core.Ants
@@ -7,12 +8,12 @@ namespace AntColony.Core.Ants
     {
         public static int Count = 45;
         public int StartPoint { get; }
-        public double Pheromone { get; }
+        public int Pheromone { get; }
         public int PathCost { get; set; }
         public List<int> Path { get; }
         public List<int> PossibleWays { get; set; }
         public List<int> BlackList { get; }
         public List<int> InitWays(int size);
-        public void Move(Graph graph, double[,] pheromones, int beta = 0, int alpha = 0);
+        public void Move(Graph graph, int[,] pheromones, int beta = 0, int alpha = 0);
     }
 }
